Don't error if storing markers in localstorage throws an error
The bug on some versions of safari was caused by this - for reasons unknown it was throwing a quota exceeded error when we go to cache the markers.
The app appears to work without this in any case, so we just catch the error and log an informational message before carrying on, which fixes the safari issue.